Use `this-command` to distinguish between a yank or delete in `evil-kill-new`
The variable `evil-is-yank-and-delete` was used to
determine if a yank command was going to be followed by a
deletion of text from the buffer so that the copied text
could be stored in either the yank or delete register.

We can simply use `this-command` set to `evil-delete` to
determine if the text should be put in the yank or delete
register. It makes more sense to do it this way because
determining if the current command should be interpreted as
a text deletion command is exactly the purpose of
`evil-is-yank-and-delete`. It is also more in-line with how
`this-command` is used in other parts of `evil`.

If other commands that aren't `evil-delete` would like to
tell `evil-yank` that the yanked text should be put in the
delete register, they just let bind `this-command` to
`evil-delete`.